broke data classes out into their own files
