+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data/ControllerInputActionOption.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a1857d39eeeb034fb61982a68634e34
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
