+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data/ControllerInputActionOptions.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.﻿
+
+using System;
+using System.Collections.Generic;
+
+namespace Microsoft.MixedReality.Toolkit.Inspectors.Data
+{
+    [Serializable]
+    public class ControllerInputActionOptions
+    {
+        public List<ControllerInputActionOption> Controllers;
+    }
+}
