+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data/ControllerInputActionOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9517fcc8e5ebbbb48a9f9b15d57752c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
